# submission/openapi.yaml
openapi: "3.0.3"
info:
  title: HR Management System API
  version: "2.0.0"
  description: |
    REST API for the Stella Technology HR Management System.

    ## Roles
    | Role | Access |
    |------|--------|
    | **ADMIN** | Registers employees & departments, views all leaves and payroll (read-only on those), approves HR employees' leaves |
    | **HR** | Manages their department only — approves employee leaves, generates payroll; submits own leaves (approved by Admin) |
    | **EMPLOYEE** | Applies for leave, views own leaves, payroll, and profile |

    ## Authentication
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click **Authorize** → enter `Bearer <token>`
    4. All subsequent requests include the token automatically

  contact:
    name: Hassan Raheem
    email: hmarwat49@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LeaveRequest:
      type: object
      required: [leaveType, startDate, endDate]
      properties:
        leaveType: { type: string, enum: [ANNUAL, CASUAL, SICK, UNPAID] }
        startDate: { type: string, format: date, example: "2026-10-01" }
        endDate:   { type: string, format: date, example: "2026-10-05" }
        reason:    { type: string, example: "Family vacation" }
    LeaveStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [APPROVED, REJECTED] }
    LeaveResponse:
      type: object
      properties:
        leaveId:    { type: string,  example: "LVE-0001" }
        employeeId: { type: string,  example: "EMP-0002" }
        leaveType:  { type: string,  enum: [ANNUAL, CASUAL, SICK, UNPAID] }
        startDate:  { type: string,  format: date }
        endDate:    { type: string,  format: date }
        totalDays:  { type: integer, example: 5 }
        reason:     { type: string,  nullable: true }
        status:     { type: string,  enum: [PENDING, APPROVED, REJECTED] }
        appliedAt:  { type: string,  format: date-time }
        updatedAt:  { type: string,  format: date-time }
    LeaveBalanceResponse:
      type: object
      properties:
        employeeId:   { type: string,  example: "EMP-0001" }
        annualLeaves: { type: integer, example: 18, description: "Default: 20. Decremented on approval." }
        casualLeaves: { type: integer, example: 10, description: "Default: 10." }
        sickLeaves:   { type: integer, example: 8,  description: "Default: 15." }
        unpaidLeaves: { type: integer, example: 30, description: "Default: 30." }

    # ── Payroll ───────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:    { type: integer, example: 400 }
        message:   { type: string,  example: "Validation failed" }
        timestamp: { type: string,  format: date-time }
  responses:
    Unauthorized:
      description: "401 — Missing or expired JWT token"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: "403 — Authenticated but role does not have permission"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: "404 — Resource not found"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: "400 — Validation error or business rule violation"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - BearerAuth: []
...
paths:
...
  /api/leaves:
    post:
      tags: [Leaves]
      summary: Apply for leave
      description: |
        Submits a leave request. The `employeeId` is automatically derived from the JWT — it does not need to be in the request body.
        - **HR employees**: their leave goes to **Admin** for approval (HR cannot approve their own type).
        - **EMPLOYEE**: their leave goes to HR for approval.
        Accessible by ADMIN, HR, and EMPLOYEE.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaveRequest' }
      responses:
        "201":
          description: Leave request submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }

    get:
      tags: [Leaves]
      summary: List leave requests
      description: |
        Returns leave requests based on the caller's role:
        - **ADMIN**: all leave requests org-wide (including HR employees' leaves).
        - **HR**: only leave requests from their own department (auto-scoped by JWT). HR employees' own leaves do not appear here.
        Optional `status` filter applies for both roles.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [PENDING, APPROVED, REJECTED] }
          description: Filter by leave status
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LeaveResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/leaves/{leaveId}/status:
    put:
      tags: [Leaves]
      summary: Approve or reject a leave request
      description: |
        Updates the status of a PENDING leave request to APPROVED or REJECTED.
        - **HR**: can approve/reject leaves for non-HR employees in their department.
          Returns **403** if the leave belongs to an HR-type employee.
        - **ADMIN**: can approve/reject any leave, including HR employees' leaves.
        Only PENDING requests can be actioned — returns 400 for already-processed requests.
        Approving an ANNUAL/CASUAL/SICK/UNPAID leave automatically decrements the employee's balance.
      security:
        - BearerAuth: []
      parameters:
        - name: leaveId
          in: path
          required: true
          schema: { type: string, example: "LVE-0001" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaveStatusRequest' }
      responses:
        "200":
          description: Leave status updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveResponse' }
        "400":
          description: Leave is not PENDING, or validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403":
          description: "403 — HR attempting to action an HR employee's leave"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/leaves/balance:
    get:
      tags: [Leaves]
      summary: List leave balances
      description: |
        Returns leave balances based on the caller's role:
        - **ADMIN**: all employees, with optional `name` filter.
        - **HR**: only employees in their own department (auto-scoped).
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          schema: { type: string, example: "John" }
          description: Filter by employee name (ADMIN only — ignored for HR)
      responses:
        "200":
          description: Leave balance list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LeaveBalanceResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/leaves/balance/{employeeId}:
    get:
      tags: [Leaves]
      summary: Get leave balance for a specific employee
      description: Returns the remaining leave balance for one employee. Accessible by ADMIN, HR, and the employee themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveBalanceResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════
  #  PAYROLL
  # ════════════════════════════════════════════════════════════
...